REF: move heavier imports to *_main.py files, allow subcommand argparsers to be collected independently
